Refactor: include validators and serializermethods in MediaUploadMixin
